did my homework... datepicker for birth in zooAdmin and contextual menu in zooAdmin ...and secured zooAdmin

+++ wpfzoo/viewmodel/ZooAdminVM.cs
@@ -38,6 +38,7 @@ namespace wpfzoo.viewmodel
         private void InitUC()
         {
             currentZoo = new Zoo();
+            currentZoo.Birth = DateTime.Now;
             this.zooAdmin.ucZoo.Zoo = currentZoo;
         }
 
@@ -46,27 +47,26 @@ namespace wpfzoo.viewmodel
             this.zooAdmin.btnValidateZoo.Click += BtnValidate_Click;
             this.zooAdmin.btnDelZoo.Click += BtnDel_Click;
             this.zooAdmin.btnNewZoo.Click += BtnNew_Click;
-            //this.zooadmin.btnaddress.click += btnaddress_click;
-            //this.zooadmin.btnemployee.click += btnemployee_click;
-            //this.zooadmin.btnstructure.click += btnstructure_click;
+            this.zooAdmin.UCZooList.DuplicateZooContextMenu.Click += DuplicateZoo_Click;
+            this.zooAdmin.UCZooList.RemoveZooContextMenu.Click += BtnDel_Click;
             this.zooAdmin.UCZooList.ItemsList.SelectionChanged += ItemsList_SelectionChanged;
         }
 
-        //private void BtnStructure_Click(object sender, RoutedEventArgs e)
-        //{
-            
-        //    throw new NotImplementedException();
-        //}
 
-        //private void BtnEmployee_Click(object sender, RoutedEventArgs e)
-        //{
-        //    throw new NotImplementedException();
-        //}
+        private void DuplicateZoo_Click(object sender, RoutedEventArgs e)
+        {
+            Zoo zoo = new entities.Zoo();
+            zoo.Birth = this.zooAdmin.ucZoo.Zoo.Birth;
+            zoo.Address = this.zooAdmin.ucZoo.Zoo.Address;
+            zoo.Name = this.zooAdmin.ucZoo.Zoo.Name;
+            zoo.Staff = this.zooAdmin.ucZoo.Zoo.Staff;
+            zoo.Structures = this.zooAdmin.ucZoo.Zoo.Structures;
 
-        //private void BtnAddress_Click(object sender, RoutedEventArgs e)
-        //{
-        //    throw new NotImplementedException();
-        //}
+            Task <Zoo> tZoo = zooManager.Insert(zoo);
+            Zoo zooRes = (Zoo)tZoo.Result;
+            this.zooAdmin.ucZoo.Zoo = zooRes;
+            InitLUC();
+        }
 
         private void ItemsList_SelectionChanged(object sender, SelectionChangedEventArgs e)
         {
@@ -79,30 +79,49 @@ namespace wpfzoo.viewmodel
 
         private async void BtnValidate_Click(object sender, System.Windows.RoutedEventArgs e)
         {
-            if (this.zooAdmin.ucZoo.Zoo.Id != 0)
+            if (this.zooAdmin.ucZoo.Zoo.Birth < DateTime.Now)
             {
-                await zooManager.Update(this.zooAdmin.ucZoo.Zoo);
+                if (this.zooAdmin.ucZoo.Zoo.Id != 0)
+                {
+                    await zooManager.Update(this.zooAdmin.ucZoo.Zoo);
+                }
+                else
+                {
+                    Task<Zoo> tZoo = zooManager.Insert(this.zooAdmin.ucZoo.Zoo);
+                    Zoo zoo = (Zoo)tZoo.Result;
+                    this.zooAdmin.ucZoo.Zoo = zoo;
+                    InitLUC();
+                }
             }
             else
             {
-                Task<Zoo> tZoo = zooManager.Insert(this.zooAdmin.ucZoo.Zoo);
-                Zoo zoo = (Zoo)tZoo.Result;
-                this.zooAdmin.ucZoo.Zoo = zoo;
-                InitLUC();
-            }
+                System.Windows.Forms.MessageBox.Show("Please do not create a zoo in the future... ", "Delete Zoo",
+                 MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
 
+                currentZoo.Birth = DateTime.Now;
+            }
+            
         }
 
         private async void BtnDel_Click(object sender, RoutedEventArgs e)
         {
-            if (System.Windows.Forms.MessageBox.Show("Do you want to delete "+ this.zooAdmin.ucZoo.Zoo.Name + " Zoo?", "Delete Zoo",
-         MessageBoxButtons.YesNo, MessageBoxIcon.Question)
-         == DialogResult.Yes)
+            if (this.zooAdmin.ucZoo.Zoo.Id != 0)
             {
-                this.zooAdmin.UCZooList.Obs.Remove(zooAdmin.ucZoo.Zoo);
-                await zooManager.Delete(zooAdmin.ucZoo.Zoo);
-                currentZoo = new Zoo();
-                this.zooAdmin.ucZoo.Zoo = currentZoo;
+                if (System.Windows.Forms.MessageBox.Show("Do you want to delete " + this.zooAdmin.ucZoo.Zoo.Name + " Zoo?", "Delete Zoo",
+                 MessageBoxButtons.YesNo, MessageBoxIcon.Question)
+                 == DialogResult.Yes)
+                {
+                    this.zooAdmin.UCZooList.Obs.Remove(zooAdmin.ucZoo.Zoo);
+                    await zooManager.Delete(zooAdmin.ucZoo.Zoo);
+                    currentZoo = new Zoo();
+                    this.zooAdmin.ucZoo.Zoo = currentZoo;
+                }
+            }
+            else
+            {
+                System.Windows.Forms.MessageBox.Show("You must select an object to delete it... ", "Delete Zoo",
+                 MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+                 
             }
         }
